backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic[email]==2.5.3
python-multipart==0.0.6

# WebSocket
websockets==12.0

# CORS
python-dotenv==1.0.0

# Trading Bot Dependencies (기존 requirements.txt와 동일)
pyupbit==0.2.35
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2

# Database


# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging
python-json-logger==2.0.7

# HTTP Client
httpx==0.26.0
